Refactor UIImageRenderer and ImageClip to use float for dimensions
- Updated UIImageRenderer to maintain sub-pixel precision for width and height calculations, improving keyframe interpolation.
- Changed ImageClip to pass float values for width and height in the updateLayout method.
- Modified ImageOverlay to store width and height as floats for better accuracy in rendering.

+++ src/main/java/mchorse/bbs_mod/camera/clips/misc/ImageOverlay.java
@@ -21,8 +21,8 @@ public class ImageOverlay
 
     public int x;
     public int y;
-    public int width;
-    public int height;
+    public float width;
+    public float height;
     public float anchorX;
     public float anchorY;
     public float windowX;
@@ -45,7 +45,7 @@ public class ImageOverlay
         this.textureBlend = textureBlend;
     }
 
-    public void updateLayout(int x, int y, int width, int height, float anchorX, float anchorY, float windowX, float windowY, float opacity)
+    public void updateLayout(int x, int y, float width, float height, float anchorX, float anchorY, float windowX, float windowY, float opacity)
     {
         this.x = x;
         this.y = y;
